only delete old incomplete files if this is the initial purge (from starting limewire up).  otherwise, files currently being downloaded will be deleted [or removed from incomplete file manager's memory, depending on if the filesystem allows files to be deleted while others are accessing them] causing strange corruptions.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/downloader/IncompleteFileManager.java
@@ -99,7 +99,7 @@ public class IncompleteFileManager implements Serializable {
         //Remove any blocks for which the file doesn't exist.
         for (Iterator iter=blocks.keySet().iterator(); iter.hasNext(); ) {
             File file=(File)iter.next();
-            if (!file.exists() || isOld(file)) {
+            if (!file.exists() || (isOld(file) && initialPurge) ) {
                 ret=true;
                 RouterService.getFileManager().removeFileIfShared(file);
                 file.delete();  //always safe to call; return value ignored
